requirements.txt
django-filter
Pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
setuptools
requests
python-decouple
python-dotenv
django-storages